+++ src/app/features/tenant/create-tenant/create-tenant.component.ts
@@ -3,6 +3,8 @@ import {FormBuilder, Validators} from "@angular/forms";
 import {BreakpointObserver} from "@angular/cdk/layout";
 import {map, Observable} from "rxjs";
 import {StepperOrientation} from "@angular/material/stepper";
+import {TenantService} from "../services/tenant.service";
+import {TenantBusinessModel} from "../../../common/business-models/tenant.business-model";
 
 @Component({
   selector: 'app-create-tenant',
@@ -16,11 +18,11 @@ export class CreateTenantComponent implements OnInit {
     birthday: [''],
   });
   postalAddressFormGroup = this._formBuilder.group({
-    road: ['', Validators.required],
+    street: ['', Validators.required],
     additionalAddress: [''],
     town: ['', Validators.required],
     // Matches 5 numeric digits, such as a zip code :
-    zip: ['', [Validators.pattern('^\\d{5}$')]],
+    zipCode: ['', [Validators.pattern('^\\d{5}$')]],
   });
   contactDetailsFormGroup = this._formBuilder.group({
     // Matches 99.99% of e-mail addresses (excludes IP e-mails)
@@ -34,6 +36,7 @@ export class CreateTenantComponent implements OnInit {
 
   constructor(
     private _formBuilder: FormBuilder,
+    private tenantService: TenantService,
     breakpointObserver: BreakpointObserver,
   ) {
     this.stepperOrientation = breakpointObserver
@@ -45,11 +48,7 @@ export class CreateTenantComponent implements OnInit {
   ngOnInit(): void {
   }
 
-  public submit(): void {
-    const identityFormData = this.identityFormGroup.value;
-    const postalAddressFormData = this.postalAddressFormGroup.value;
-    const contactDetailsFormData = this.contactDetailsFormGroup.value;
-    const birthday = identityFormData['birthday'];
-    console.warn(birthday);
+  public onSubmit($event: TenantBusinessModel): void {
+    this.tenantService.postTenant($event).subscribe();
   }
 }