+++ src/app/features/tenant/tenant-detail/tenant-detail.component.ts
@@ -1,5 +1,6 @@
 import {Component, OnInit} from '@angular/core';
-import {ActivatedRoute} from "@angular/router";
+import {ActivatedRoute, Router} from "@angular/router";
+import {TenantService} from "../services/tenant.service";
 
 @Component({
   selector: 'app-tenant-detail',
@@ -10,10 +11,23 @@ export class TenantDetailComponent implements OnInit {
 
   public tenantID!: string;
 
-  constructor(private activatedRoute: ActivatedRoute) {
+  constructor(private activatedRoute: ActivatedRoute,
+              private router: Router,
+              private tenantService: TenantService) {
   }
 
   ngOnInit(): void {
     this.tenantID = this.activatedRoute.snapshot.params['id_tenant'];
+  }
+
+  public delete(): void {
+    this.tenantService.deleteTenant(this.tenantID).subscribe({
+      next: () => {
+        this.router.navigate(["/locataires"]);
+      },
+      error: () => {
+        // TODO add toast
+      }
+    })
   }
 }